tests: separate common API into sharing files
